refactor(tests): move kafka identity tests to dedicated file
Move 11 identity validation tests from test_host_mq_service.py
to test_mq_kafka_identity.py for better organization.
